fix relations

+++ db/migrate/20161203215306_create_trips.rb
@@ -2,8 +2,8 @@ class CreateTrips < ActiveRecord::Migration[5.0]
   def change
     create_table :trips do |t|
       t.bigint    :duration
-      t.bigint    :start_station_id
-      t.bigint    :end_station_id
+      t.integer    :start_station_id
+      t.integer    :end_station_id
       t.datetime  :start_date_time
       t.date      :start_date
       t.datetime  :end_date_time
